refactor(RVCV-71): ReviewDesign 컨트롤러단 클래스를 분리한다
