+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ReviewDesignUseCase.java
@@ -1,0 +1,8 @@
+package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
+
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.UpdateReviewDesignRequest;
+
+public interface ReviewDesignUseCase {
+	void updateReviewDesign(Integer adminId, Integer reviewDesignId,
+		UpdateReviewDesignRequest updateReviewDesignRequest);
+}
