+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCase.java
@@ -1,12 +1,12 @@
 package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
 
-import com.romanticpipe.reviewcanvas.domain.ReviewDesign;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.SignUpRequest;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.UpdateReviewDesignRequest;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.response.GetReviewVisibilityTitleResponse;
+import java.util.List;
+
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
+import com.romanticpipe.reviewcanvas.domain.ReviewDesign;
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.SignUpRequest;
+import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.response.GetReviewVisibilityTitleResponse;
 
 public interface ShopAdminUseCase {
 
@@ -18,6 +18,4 @@ public interface ShopAdminUseCase {
 
 	List<ReviewDesign> getGeneralReviewThemeList();
 
-	void updateReviewDesign(Integer adminId, Integer reviewDesignId,
-							UpdateReviewDesignRequest updateReviewDesignRequest);
 }
