+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCaseImpl.java
@@ -1,27 +1,26 @@
 package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
 
+import java.util.List;
+
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
 import com.romanticpipe.reviewcanvas.domain.AdminAuth;
 import com.romanticpipe.reviewcanvas.domain.ReviewDesign;
 import com.romanticpipe.reviewcanvas.domain.ReviewDesignType;
 import com.romanticpipe.reviewcanvas.domain.ReviewVisibility;
 import com.romanticpipe.reviewcanvas.domain.ShopAdmin;
 import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.SignUpRequest;
-import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.request.UpdateReviewDesignRequest;
 import com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase.response.GetReviewVisibilityTitleResponse;
 import com.romanticpipe.reviewcanvas.service.AdminAuthCreater;
-import com.romanticpipe.reviewcanvas.service.MyReviewDesignValidator;
 import com.romanticpipe.reviewcanvas.service.ReviewDesignReader;
-import com.romanticpipe.reviewcanvas.service.ReviewDesignValidator;
 import com.romanticpipe.reviewcanvas.service.ReviewVisibilityReader;
 import com.romanticpipe.reviewcanvas.service.ShopAdminCreator;
 import com.romanticpipe.reviewcanvas.service.ShopAdminValidator;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
+import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
@@ -33,8 +32,6 @@ class ShopAdminUseCaseImpl implements ShopAdminUseCase {
 	private final ShopAdminValidator shopAdminValidator;
 	private final ReviewVisibilityReader reviewVisibilityReader;
 	private final ReviewDesignReader reviewDesignReader;
-	private final ReviewDesignValidator reviewDesignValidator;
-	private final MyReviewDesignValidator myReviewDesignValidator;
 
 	@Override
 	@Transactional
@@ -85,30 +82,4 @@ class ShopAdminUseCaseImpl implements ShopAdminUseCase {
 		return reviewDesignReader.getGeneralThemeList(ReviewDesignType.GENERAL);
 	}
 
-	@Override
-	@Transactional
-	public void updateReviewDesign(Integer shopAdminId, Integer reviewDesignId,
-								   UpdateReviewDesignRequest updateReviewDesignRequest) {
-		ReviewDesign reviewDesign = reviewDesignValidator.validById(reviewDesignId);
-		myReviewDesignValidator.validateIsMyDesign(shopAdminId, reviewDesignId);
-
-		reviewDesign.update(
-			updateReviewDesignRequest.reviewDesignPosition(),
-			updateReviewDesignRequest.themeName(),
-			updateReviewDesignRequest.layoutType(),
-			updateReviewDesignRequest.padding(),
-			updateReviewDesignRequest.gap(),
-			updateReviewDesignRequest.boxShadowColor(),
-			updateReviewDesignRequest.boxShadowWidth(),
-			updateReviewDesignRequest.borderColor(),
-			updateReviewDesignRequest.borderTransparency(),
-			updateReviewDesignRequest.borderWidth(),
-			updateReviewDesignRequest.pagingType(),
-			updateReviewDesignRequest.pagingNumber(),
-			updateReviewDesignRequest.textAlign(),
-			updateReviewDesignRequest.pointColor(),
-			updateReviewDesignRequest.pointType(),
-			updateReviewDesignRequest.lineEllipsis(),
-			updateReviewDesignRequest.reviewDesignUrl());
-	}
 }
